+++ time/src/most.ts
@@ -2,10 +2,10 @@ import * as most from 'most';
 import {Stream} from 'most';
 import {setAdapt} from '@cycle/run/lib/adapt';
 
-import {mockTimeSource as mockTimeSourceUntyped} from './src/mock-time-source';
-import {timeDriver as timeDriverUntyped} from './src/time-driver';
-import {Frame} from './lib/cjs/src/animation-frames';
-import {Comparator, OperatorArgs} from './src/types';
+import {mockTimeSource as mockTimeSourceUntyped} from './mock-time-source';
+import {timeDriver as timeDriverUntyped} from './time-driver';
+import {Frame} from './animation-frames';
+import {Comparator, OperatorArgs} from './types';
 
 setAdapt(stream => most.from(stream as any));
